moving BasicWebChat CSS logic (removing unused)
