+++ packages/component/src/BasicTranscript.tsx
@@ -93,6 +93,10 @@ const ROOT_STYLE = {
   }
 };
 
+const TRANSCRIPT_STYLE = {
+  flex: 1
+};
+
 type RenderingElement = {
   activity: WebChatActivity;
   callbackRef: (element: HTMLElement) => void;
@@ -132,7 +136,9 @@ const InternalTranscript = forwardRef<HTMLDivElement, InternalTranscriptProps>(
     const getKeyByActivity = useGetKeyByActivity();
     const getKeyByActivityId = useGetKeyByActivityId();
     const localize = useLocalizer();
-    const rootClassName = useStyleToEmotionObject()(ROOT_STYLE) + '';
+    const styleToEmotionObject = useStyleToEmotionObject();
+    const rootClassName = styleToEmotionObject(ROOT_STYLE) + '';
+    const transcriptClassName = styleToEmotionObject(TRANSCRIPT_STYLE) + '';
     const rootElementRef = useRef<HTMLDivElement>();
     const terminatorLabelId = useUniqueId('webchat__basic-transcript__terminator-label');
     const terminatorRef = useRef<HTMLDivElement>();
@@ -518,6 +524,7 @@ const InternalTranscript = forwardRef<HTMLDivElement, InternalTranscriptProps>(
           'webchat__basic-transcript',
           basicTranscriptStyleSet + '',
           rootClassName,
+          transcriptClassName,
           (className || '') + ''
         )}
         dir={direction}
